Move KeyedDataProvider to CLDR crate (see #442)
