+++ provider/cldr/src/transform/datetime/skeletons.rs
@@ -9,8 +9,9 @@ use crate::cldr_serde;
 use crate::CldrPaths;
 use icu_datetime::{provider::*, skeleton::SkeletonError};
 
+use crate::support::KeyedDataProvider;
 use icu_plurals::PluralCategory;
-use icu_provider::iter::{IterableDataProviderCore, KeyedDataProvider};
+use icu_provider::iter::IterableDataProviderCore;
 use icu_provider::prelude::*;
 use std::convert::TryFrom;
