+++ provider/cldr/src/transform/list/mod.rs
@@ -5,11 +5,12 @@
 use crate::cldr_serde;
 use crate::error::Error;
 use crate::reader::{get_langid_subdirectories, open_reader};
+use crate::support::KeyedDataProvider;
 use crate::CldrPaths;
 use icu_list::provider::*;
 use icu_locid::LanguageIdentifier;
 use icu_locid_macros::langid;
-use icu_provider::iter::{IterableDataProviderCore, KeyedDataProvider};
+use icu_provider::iter::IterableDataProviderCore;
 use icu_provider::prelude::*;
 use litemap::LiteMap;
 use std::convert::TryFrom;
