+++ provider/cldr/src/transform/time_zones/mod.rs
@@ -5,13 +5,12 @@
 use crate::cldr_serde;
 use crate::error::Error;
 use crate::reader::{get_langid_subdirectories, open_reader};
+use crate::support::KeyedDataProvider;
 use crate::CldrPaths;
 use icu_datetime::provider::{key, time_zones::*};
 use icu_locid::LanguageIdentifier;
-use icu_provider::{
-    iter::{IterableDataProviderCore, KeyedDataProvider},
-    prelude::*,
-};
+use icu_provider::iter::IterableDataProviderCore;
+use icu_provider::prelude::*;
 use litemap::LiteMap;
 
 use std::convert::TryFrom;
