+++ tools/datagen/src/bin/datagen.rs
@@ -6,11 +6,11 @@ use clap::{App, Arg, ArgGroup, ArgMatches};
 use eyre::WrapErr;
 use icu_locid::LanguageIdentifier;
 use icu_properties::provider::key::{ALL_MAP_KEYS, ALL_SET_KEYS};
+use icu_provider::either::EitherProvider;
 use icu_provider::export::DataExporter;
 use icu_provider::filter::Filterable;
 use icu_provider::hello_world::{self, HelloWorldProvider};
 use icu_provider::prelude::*;
-use icu_provider::either::EitherProvider;
 use icu_provider::serde::SerializeMarker;
 use icu_provider_blob::export::BlobExporter;
 use icu_provider_cldr::download::CldrAllInOneDownloader;
